Refactoring: service.go now concentrates all stuff related to Skynet Services. RemoveFromConfig and AddToConfig now takes a Service as arg. RegisterHeartbeat moved to config.go leaving types.go with just types.
config.go/GetRandomClientByProvides received some cleanner var names.
